Limit hitscan attack distance to `INT_MAX`
Something not right. Range isn`t working as LONG_MAX, despite of all necessary variables has been updated to int64 types.

For now, probably better to limit range to INT_MAX (which is how far range actually is) to avoid possible overflows.

+++ src/doom/p_map.c
@@ -1545,7 +1545,7 @@ void P_LineAttack (mobj_t *t1, angle_t angle, int64_t distance, fixed_t slope, i
     // (see PTR_ShootTraverse).
     if (distance >= MISSILERANGE && singleplayer && !vanillaparm)
     {
-        distance = LONG_MAX;
+        distance = INT_MAX;
     }
 
     angle >>= ANGLETOFINESHIFT;
